feat(refactor): create pool accounts to hooks (#1913)
