+++ src/contexts/Pools/FavoritePools/defaults.ts
@@ -2,14 +2,13 @@
 // SPDX-License-Identifier: GPL-3.0-only
 /* eslint-disable @typescript-eslint/no-unused-vars, @typescript-eslint/no-empty-function */
 
-import type { PoolsConfigContextState } from './types';
+import type { FavoritePoolsContextState } from './types';
 import type { PoolAddresses } from '../BondedPools/types';
 
-export const defaultPoolsConfigContext: PoolsConfigContextState = {
-  addFavorite: () => {},
-  removeFavorite: () => {},
-  createAccounts: () => poolAddresses,
+export const defaultFavoritePoolsContext: FavoritePoolsContextState = {
   favorites: [],
+  addFavorite: (address: string) => {},
+  removeFavorite: (address: string) => {},
 };
 
 export const poolAddresses: PoolAddresses = {
