+++ src/contexts/Pools/FavoritePools/types.ts
@@ -1,0 +1,8 @@
+// Copyright 2023 @paritytech/polkadot-staking-dashboard authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+export interface FavoritePoolsContextState {
+  favorites: string[];
+  addFavorite: (address: string) => void;
+  removeFavorite: (address: string) => void;
+}
