--- src/contexts/Pools/PoolsConfig/types.ts
@@ -1,11 +1,0 @@
-// Copyright 2023 @paritytech/polkadot-staking-dashboard authors & contributors
-// SPDX-License-Identifier: GPL-3.0-only
-
-import type { PoolAddresses } from '../BondedPools/types';
-
-export interface PoolsConfigContextState {
-  addFavorite: (a: string) => void;
-  removeFavorite: (a: string) => void;
-  createAccounts: (p: number) => PoolAddresses;
-  favorites: string[];
-}
